refactor : gameController에서 RacingGame 분리
